pyproject.toml
[project]
name = "ai-agent-prototype"
version = "0.1.0"
description = "IA Agent with tool calling and memory."
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "openai>=1.93.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
[project.optional-dependencies]
dev = ["ruff", "pyright", "pytest", "pytest-xdist"]
# ============================
# Lint e formatação (Ruff)
# ============================
[tool.ruff]
line-length = 88
target-version = "py313"
fix = true
show-fixes = true
indent-width = 4
exclude = ["venv", ".venv", "env", ".env", "node_modules", "__pycache__"]

[tool.ruff.lint]
select = [
    "ASYNC",
    "A",
    "ANN",
    "B",
    "BLE",
    "C4",
    "C90",
    "COM",
    "E",
    "EM",
    "ERA",
    "EXE",
    "F",
    "FBT",
    "FIX",
    "I",
    "ICN",
    "ISC",
    "Q",
    "RET",
    "RSE",
    "S",
    "SIM",
    "SLF",
    "T10",
    "T20",
    "TC",
    "TD",
    "TRY",
    "UP",
    "W",
    "YTT",
    "RUF",
    "N",
]
ignore = ["T201", "COM812"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ANN201", "S101"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.ruff.lint.isort]
# Ensina ao Ruff qual é o pacote principal do seu projeto.
# Deve ser o mesmo nome da pasta dentro de 'src/'.
known-first-party = ["my_package"]


# ============================
# Tipagem (Pyright)
# ============================
[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.13"
include = ["src", 'tests']
exclude = [
    "**/venv",
    "**/.venv",
    "**/env",
    "**/.env",
    "**/node_modules",
    "**/__pycache__",
]
venv = ".venv"
venvPath = "."
executionEnvironments = [{ root = "src" }]

# ============================
# Testes (Pytest)
# ============================
[tool.pytest.ini_options]
addopts = "-s --color=yes --tb=short"
# Adiciona 'src' ao path do Python durante os testes. Isso permite que
# os testes em `tests/` importem os módulos de `src/` sem erros.
pythonpath = ["src"]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "ruff>=0.12.2",
]
